consolidated the playlist overview into one page
