--- frontend/pages/Playlist/MyPlaylists.tsx
@@ -1,14 +1,0 @@
-import { Box, Center, Flex } from "@chakra-ui/react";
-import Navbar from "../../components/shared/Navbar";
-import PlaylistSidebar from "../../components/playlist/PlaylistSidebar";
-
-export default function MyPlaylistsPage() {
-  return (
-    <>
-      <Navbar />
-      <Flex>
-        <PlaylistSidebar />
-      </Flex>
-    </>
-  );
-}
